enabled electric-pairs and changed flyspell keybindings
Enabled electric-pairs mode to have brackets and such paired by
default, then created a function to have them work with highlighted
text in org-mode so I can bold or italic existing text easier when
I want to. Also changed the flyspell bindings to F8 with mods so
it's easier to remember and use.