switched to Emacs version repos and doom modeline
I had a feeling some of the packages I was using were likely for a newer
version of Emacs, so I switched to it using the version number to decide
which repo to use. That cleaned up the packages considerably.
Also switched to Doom modeline and configured it. I like the cleaner look.