Cleaned up some of the cruft from customize
When I changed the theme, customize added a lot of color options
to the config that I don't want or need. Removed them.
Also changed key binding for flyspell, to get the word at point
to be the one with spelling suggestions. Had the wrong binding
originally so was getting wrong suggestions.