Removed all the comments and reset garbage collection
The comments are really unnecessary for me now, so I removed them to reduce
the file lines and size. Tried setting the garbage collection to what is the
default for Spacemacs, but I get the stuttering back. So I set what I think
will be comfortable, 20Mb. It does start a little slower, but I don't have the
stuttering when I'm working, so I'll live with the slower start.