added code to make helm not size to split window
Didn't like how Helm only showed in window split it was called from
so I found some code from one of the Spacemacs devs on Reddit that
made Helm use the whole window width as it does in Spacemacs. That
allows me to see more of search matches with less scrolling.