added helm-resume keybinding
Added a key binding to make Helm resume the last search. Mostly for
use with ripgrep, so I can go back to what I was searching for if
the result I picked isn't what I was really looking for, or if I'm
making changes and want to find the next one to change.