change some of the helm defaults to show bindings in split
Haven't tried it yet, but changed what I hope is the default to
allow helm to show it's keybindings when I do a C-h m so I can
see what keystrokes are available. By default it won't show them
due to my having the helm window span the entire workspace.