Removed all the Python stuff
It's too much of a pain in the ass for me to try and learn
Python and at the same time have to try and figure out how
to get the editor to work properly with it. So I just got
rid of it and will stick with VSCode for now.