removed pdf-tool due to bug
pdf-tools worked great, very efficient and fast, but had bad memory
bug that caused it to eat a lot of memory and not release it. Read
bug reports on the issue, seems Emacs itself doesn't work with
memory in a way that makes it easy for pdf-tools to release it, so
the bug goes unresolved.