Set the default dictionary for flyspell
Made a few usability changes, the main one to flyspell for finding
the default dictionary. Remapped the key to enable it, and pointed
it to the US dictionary. Also removed the menu bar and changed the
cursor back to a block, the other was just too hard to see even when
flashing.