Removed the tasklist widgets, don't need them
The tasklist was more an annoyance than useful. I use a border color
to tell which window is active, and never minimize, so there was no
need for a tasklist. Removed the code for the widgets. Added a textbox
spacer to keep the top bar balanced between left and right widgets.