Start the server from first instance
Considering the daemon doesn't read the fonts properly and would
require re-evaluating the font settings to work, I decided to just
start the server when I open the first session. Any future clients
will then have the proper font setting. Apparently it's a problem
with the daemon not being attached to anything prior to creating
the frame.