Switched back to Ivy. Found CPU problem from Helm.
Noticed the CPU was constantly in use, even when Emacs was idle and not in
use. Doing a profile showed Helm was using CPU when doing nothing at all.
Switched back to Ivy, and CPU is nil when not actively using it. Apparently a
bug in Helm that isn't likely to get fixed.