Removed Evil mode, causing too many problems
Though I like the idea of Evil being like Vim, it adds to many odd behaviors to some
of the other packages, like Roam. Spend too much time trying to find and fix problems,
it's easier to stick with Emacs bindings as the other packages are meant to be used.