Added Evil mode back, think I found problem
I believe some of the problem I had with it before was not really knowing how to
set it up, how it works. Giving it another try to see if I can keep it workable. Do
miss having the easier key strokes of Vim and a leader key, faster and easier
than always having to use CTRL.