Tried to get LSP to work like it does in Neovim
I know I'm missing something, but no matter what I try I can't get
LSP mode in Emacs to work as well as it does in Neovim. With the
Typescript libraries installed, it still can't find where a function
or method is described, and seems to insist on showing what it can
find in two different places, the minibuffer and another frame.