Removed Roam and friends
While a great piece of software, I find it more work to use it
properly, certainly more than I need. Anything I need for my notes
is already provided by Org, and a lot easier to use. So after trying
my best with it for several months, I've decided it's not for me.