Changed Evil visual line keybinding for arrows
Using the arrow keys when browsing never seemed to work right, though
it did with j & k. I suspect the arrows are not part of the normal
Evil mappings, so I moved them to a global Evil define key mapping
instead, and now they appear to be working properly.

Also re-ordered a few bindings to be alphabetically correct, easier
to find that way.