For some reason polybar isn't using one of the colors
When trying to launch polybar on the new SSD, I ran it from the terminal to debug,
and noticed it doesn't use one of the color declarations. Not sure why, and the color
is applied, but it would throw an error, so I commented out the color to see how it goes.