Removed variable-pitch from Org
While it looked nice, it was buggy and seemed to cause odd behavior in many
ways. Left the font setting for it for anything that might use it by default, but
removed the settings that affect Org documents and went back to default
fixed-pitch. Hopefully that will fix some of the odd textual quirks.