Removed lexical binding
There were some articles about the advantages of lexical binding in the init file,
but it seems to me now that it's only advantageous if you're familiar with elisp
and can configure the system to take advantage of it. I'm not, and I found that
it was likely causing more problems than it solved.