some more changes to LSP config
Having a hell of a time getting the prefix keys working for LSP,
for some reason the documented way to do it with a setq doesn't
work for me. Can't find out why, but found a way around it. As
long as it works I'm not going to worry about it.