Switched projectile to project.el
There's so much a package like projectile does that I'll never use.
Emacs 28 has a built-in project manager that's available in elpa
and works with Emacs 27, so I installed it instead. Much lighter,
and does all I need. Only had to create a function so it would see
my Notes folder as a project, something similar to .projectile.