Removed Roam and did some cleanup with Vertico and friends
Roam won't do it for me, it's not going to work for what I need.
So I removed it, and made some small changes with Vertico and
Consult, as well as changed from the JetBrains Nerd font to the
regular one. It actually looks a lot better, and I have no need
for a Nerd font in Emacs.