Added Roam back to try a few things
Want to see if it will work for just linking headings to other
other places, kind of like a wiki. Might be easier to keep separate
headings in a single file that have the same name. Assign them an
ID and use Roam to link to them via ID instead of name.