Experiment over, not going to work. Done trying.
Org-Roam would surely work for someone just doing random notes, but
for me where many filenames would be the same beyond the zettlename,
it just becomes too much to create workable search routines. When
looking for a file, even using tags, the way it's displayed makes it
hard to tell which is which. There's no stable pattern in which tag
would be shown first (or at all) and no matter how I enter them, it
seems their use is random. Tired of trying.