Removed osm package
Had hoped it would be useful for genealogy work, especially since
it had the ability to do a search for a place, but it turns out
while it works well, the maps don't render correctly, having all
kinds of blank spaces and artifacts. If it doesn't do as it's
suppose to, it's no use to me.