Removed all the unused plugins
With all the constant breaking changes with plugins, it was getting
annoying to have to constantly edit the config to adapt it to the
changes. I just need nvim to work, I don't need it to be some super
fancy code editor. So removed all the unnecessary packages and all
the lines of code used to set them.