Trying to clear Org error message with Org 9.6
The latest update has errors and won't open org file in org-mode.
Tried the suggested workaround of moving the loading to start org
early, but doesn't seem to make any difference. Only thing working
so far it to delete the .elc files from the elpa org folder, which
works but slows down org.