Until bug gets fixed, removed call to Org updated package
With the current assert error in Org 9.6 causing to to not work, not read
org files as org, I've removed the package requirement and deleted 9.6
from the installed packages. I think this should have been better tested
upstream, but so be it. There's nothing critical needed in 9.6 so no loss.