Removed variable-pitch font
Though it looked fine, there were some small quirks with having
both a variable and a fixed font showing on the same pages.
Sometimes the completion drop-down would not render correctly,
sometimes the text alignment would be off, sometimes the collapsed
heading markers would show when there was no more text past that
heading. Little things but annoying. Will try for a while with a
single fixed font to see if that cures it.