waybar: New config for hyprland
I renamed this file to isolate it so sway doesn't try to use it,
as I don't know whether it would work with sway. It's working fine
with hyprland, but I have some other issues with that wm I need
to iron out before I can make the changes permanent.