nvim: Set autochdir option true
For my use I want nvim to change the folder to whatever parent
of the file I'm working on. It's easier when I have to work on
other files in the same folder. By default it stays in whatever
folder I started nvim in, which can be a pain in the ass.