Fix modeline right margin
Seems something changed in doom modeline that caused the mode
name on the right side to go slightly beyond the right edge,
cutting off the end of the text. Found an issue on it's Github
with a fix. Removing the fringe-mode setting from my defaults
and setting the right-align edge fixed it.