Changed custom variables to use newer setop
There was a new macro introduced in Emacs 29 that is better for
setting custom variables than setq. Instead of setting once when
Emacs is started as setq does, it will reset the variable when
needed during program execution. That helps to avoid bugs that might
be introduced when changing modes, etc.